refactor(desktop): extract BrowserPaneToolbar into its own component
One component per file: the toolbar moves to its own folder and the
shared runtime-state subscription becomes hooks/useBrowserState, with
the barrel re-exporting both so callers are unchanged.
