+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/BrowserPane/hooks/useBrowserState/index.ts
@@ -1,0 +1,1 @@
+export { useBrowserState } from "./useBrowserState";
