+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/BrowserPane/hooks/useBrowserState/useBrowserState.ts
@@ -1,0 +1,15 @@
+import { useCallback, useSyncExternalStore } from "react";
+import {
+	type BrowserRuntimeState,
+	browserRuntimeRegistry,
+} from "../../browserRuntimeRegistry";
+
+export function useBrowserState(paneId: string): BrowserRuntimeState {
+	return useSyncExternalStore(
+		useCallback(
+			(cb) => browserRuntimeRegistry.onStateChange(paneId, cb),
+			[paneId],
+		),
+		useCallback(() => browserRuntimeRegistry.getState(paneId), [paneId]),
+	);
+}
