+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/BrowserPane/index.ts
@@ -1,6 +1,3 @@
-export {
-	BrowserPane,
-	BrowserPaneToolbar,
-	renderBrowserTabIcon,
-} from "./BrowserPane";
+export { BrowserPane, renderBrowserTabIcon } from "./BrowserPane";
 export { browserRuntimeRegistry } from "./browserRuntimeRegistry";
+export { BrowserPaneToolbar } from "./components/BrowserPaneToolbar";
